# OEMBackend/src/api/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"TaskCategoryDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"category",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResourceDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"startTime": {
						"type": "string"
					},
					"endTime": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OperationDto": {
				"properties": {
					"type": {
						"$ref": "#/components/schemas/TaskCategoryDto"
					},
					"startTime": {
						"type": "string"
					},
					"endTime": {
						"type": "string"
					},
					"resources": {
						"items": {
							"$ref": "#/components/schemas/ResourceDto"
						},
						"type": "array"
					},
					"payload": {}
				},
				"required": [
					"type",
					"startTime",
					"endTime",
					"resources"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OperationStatusDto": {
				"type": "string",
				"enum": [
					"Pending",
					"InProgress",
					"Completed",
					"Failed"
				]
			},
			"OperationWithStatusDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"operation": {
						"$ref": "#/components/schemas/OperationDto"
					},
					"status": {
						"$ref": "#/components/schemas/OperationStatusDto"
					}
				},
				"required": [
					"id",
					"operation",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VesselVisitExecutionStatusDto": {
				"type": "string",
				"enum": [
					"Open",
					"Closed"
				]
			},
			"VesselVisitExecutionDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"relatedVVN": {
						"type": "string"
					},
					"operationsExecuted": {
						"items": {
							"$ref": "#/components/schemas/OperationWithStatusDto"
						},
						"type": "array"
					},
					"dateOpen": {
						"type": "string",
						"format": "date-time"
					},
					"dateClosed": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"$ref": "#/components/schemas/VesselVisitExecutionStatusDto"
					},
					"createdBy": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"code",
					"relatedVVN",
					"operationsExecuted",
					"status",
					"createdBy"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Page_VesselVisitExecutionDto_": {
				"properties": {
					"pageNumber": {
						"type": "number",
						"format": "double"
					},
					"pageSize": {
						"type": "number",
						"format": "double"
					},
					"pageCount": {
						"type": "number",
						"format": "double"
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/VesselVisitExecutionDto"
						},
						"type": "array"
					}
				},
				"required": [
					"pageNumber",
					"pageSize",
					"pageCount",
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "oembackend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/vessel-visit-executions/{relatedVVN}/open": {
			"post": {
				"operationId": "OpenVesselVisitExecution",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/VesselVisitExecutionDto"
								}
							}
						}
					}
				},
				"tags": [
					"Vessel Visit Executions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "relatedVVN",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/vessel-visit-executions/{relatedVVN}/close": {
			"put": {
				"operationId": "CloseVesselVisitExecution",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/VesselVisitExecutionDto"
								}
							}
						}
					}
				},
				"tags": [
					"Vessel Visit Executions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "relatedVVN",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/vessel-visit-executions/{relatedVVN}/operations/start": {
			"put": {
				"operationId": "StartOperation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/VesselVisitExecutionDto"
								}
							}
						}
					}
				},
				"tags": [
					"Vessel Visit Executions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "relatedVVN",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/vessel-visit-executions/{relatedVVN}/operations/{operationId}/complete": {
			"put": {
				"operationId": "CompleteOperation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/VesselVisitExecutionDto"
								}
							}
						}
					}
				},
				"tags": [
					"Vessel Visit Executions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "relatedVVN",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "operationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"endTime": {
										"type": "string"
									}
								},
								"required": [
									"endTime"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/vessel-visit-executions/{relatedVVN}": {
			"get": {
				"operationId": "GetVesselVisitExecution",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/VesselVisitExecutionDto"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Vessel Visit Executions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "relatedVVN",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/vessel-visit-executions": {
			"get": {
				"operationId": "GetAllVesselVisitExecutions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Page_VesselVisitExecutionDto_"
								}
							}
						}
					}
				},
				"tags": [
					"Vessel Visit Executions"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}